Split inlay hints into modules
